requirements.txt
torch==2.2.0
numpy==1.26.0
pandas==2.2.0
scikit-learn==1.4.0
gradio==4.31.0
python-dotenv==1.0.0
openai==1.12.0